feat: add number to specialCharSet in TextMeasure
