+++ packages/vtable/__tests__/components/pivotTable-size-legend.test.ts
@@ -302,7 +302,7 @@ describe('listTable-size-legend init test', () => {
   test('listTable-size-legend getDrawRange', () => {
     const drawRange = pivotTable.getDrawRange();
     expect(drawRange.left).toBe(1);
-    expect(drawRange.right).toBe(566);
+    expect(drawRange.right).toBe(568);
     expect(drawRange.top).toBe(58);
     expect(drawRange.bottom).toBe(799);
   });
